+++ keta.js
@@ -519,9 +519,25 @@ angular.module('keta.directives.AppBar',
 											}
 										}
 									});
+
 									// use link-element to easily access url params
 									var link = document.createElement('a');
 									link.href = entryUri;
+
+									var linkProtocol =
+										link.protocol +
+										(link.protocol[link.protocol.length - 1] === ':' ?
+											'//' : '://');
+
+									var linkPort =
+										link.port !== '' && link.port !== '0' ?
+										':' + link.port : '';
+
+									// workaround for internet explorer not having "origin" property
+									var linkOrigin =
+										angular.isDefined(link.origin) ?
+											link.origin :
+										linkProtocol + link.hostname + linkPort;
 
 									if (entryUri !== null) {
 										scope.rootApp = {};
@@ -534,15 +550,16 @@ angular.module('keta.directives.AppBar',
 									}
 
 									scope.links.ALL_APPS = angular.isString(scope.links.ALL_APPS) ?
-										scope.links.ALL_APPS : link.origin + link.search + '#/applications';
+										scope.links.ALL_APPS : linkOrigin + link.search + '#/applications';
 
 									scope.links.USER_PROFILE = angular.isString(scope.links.USER_PROFILE) ?
-										scope.links.USER_PROFILE : link.origin + link.search + '#/user';
+										scope.links.USER_PROFILE : linkOrigin + link.search + '#/user';
 
 									if (!angular.isString(scope.links.ALL_ENERGY_MANAGERS)) {
-										scope.links.ALL_ENERGY_MANAGERS = link.origin + link.search
+										scope.links.ALL_ENERGY_MANAGERS = linkOrigin + link.search
 											+ '#/devices?deviceClass=com.kiwigrid.devices.em.EnergyManager';
 									}
+
 								}
 							});
 					}
@@ -1086,6 +1103,8 @@ angular.module('keta.directives.AppBar')
  *     data-column-class-callback="columnClassCallback"
  *     data-table-class-callback="tableClassCallback"
  *     data-row-class-callback="rowClassCallback"
+ *     data-search-input-width-classes="searchInputWidthClasses"
+ *     data-selector-width-classes="selectorWidthClasses"
  *     data-pager="pager"
  *     data-search="search"
  *     data-search-wait-ms="waitTime"
@@ -1259,8 +1278,7 @@ angular.module('keta.directives.AppBar')
  *             var rowClass = 'row-is-selected';
  *             if (isHeader) {
  *                 rowClass += ' header-row';
- *             }
- *             if (angular.isDefined(row.connected) && row.connected === true) {
+ *             } else if (angular.isDefined(row.connected) && row.connected === true) {
  *                 rowClass += ' connected';
  *             }
  *             return rowClass;
@@ -1270,6 +1288,12 @@ angular.module('keta.directives.AppBar')
  *         $scope.tableClassCallback = function() {
  *             return ['table-striped'];
  *         };
+ *
+ *         // bootstrap width classes to define the size of the search input
+ *         $scope.searchInputWidthClasses = 'col-xs-12 col-sm-6';
+ *
+ *         // bootstrap width classes to define the size of the selector dropdown
+ *         $scope.selectorWidthClasses = 'col-xs-12 col-sm-6 col-md-6 col-lg-6';
  *
  *         // object for pager configuration (total, limit, offset)
  *         // with this configuration object you are able to manage paging
@@ -1417,6 +1441,12 @@ angular.module('keta.directives.ExtendedTable',
 				// callback method to return class array for table
 				tableClassCallback: '=?',
 
+				// bootstrap width classes to define the size of the search input
+				searchInputWidthClasses: '=?',
+
+				// bootstrap width classes to define the size of the selector dropdown
+				selectorWidthClasses: '=?',
+
 				// object for pager configuration (total, limit, offset)
 				pager: '=?',
 
@@ -1533,6 +1563,12 @@ angular.module('keta.directives.ExtendedTable',
 				scope.tableClassCallback = scope.tableClassCallback || function() {
 					return ['table-striped'];
 				};
+
+				// bootstrap width classes for search input
+				scope.searchInputWidthClasses = scope.searchInputWidthClasses || 'col-xs-12 col-sm-4';
+
+				// bootstrap width classes for selector dropdown
+				scope.selectorWidthClasses = scope.selectorWidthClasses || 'col-xs-12 col-sm-8 col-md-8 col-lg-8';
 
 				// pager
 				var defaultPager = {};
@@ -1968,12 +2004,13 @@ angular.module('keta.directives.ExtendedTable')
 '}">' +
 '' +
 '	<div class="row" data-ng-show="!isDisabled(COMPONENTS_FILTER) || !isDisabled(COMPONENTS_SELECTOR)">' +
-'		<div class="col-xs-12 col-sm-6">' +
+'' +
+'		<div data-ng-class="::searchInputWidthClasses">' +
 '' +
 '			<!-- FILTER -->' +
 '			<div data-ng-show="!isDisabled(COMPONENTS_FILTER)">' +
 '				<div class="form-group form-inline">' +
-'					<div class="input-group col-xs-12 col-sm-8 col-md-6">' +
+'					<div class="input-group col-xs-12">' +
 '						<input type="search" class="form-control"' +
 '							placeholder="{{ getLabel(MESSAGE_KEY_PREFIX + \'_search\') }}"' +
 '							data-ng-model="search"' +
@@ -1989,7 +2026,8 @@ angular.module('keta.directives.ExtendedTable')
 '			</div>' +
 '' +
 '		</div>' +
-'		<div class="col-xs-12 col-sm-6 col-md-6 col-lg-6">' +
+'' +
+'		<div data-ng-class="::selectorWidthClasses">' +
 '' +
 '			<!-- SELECTOR -->' +
 '			<div data-ng-show="!isDisabled(COMPONENTS_SELECTOR) && selectedColumn !== null" class="form-horizontal">' +
@@ -2027,6 +2065,7 @@ angular.module('keta.directives.ExtendedTable')
 '			</div>' +
 '' +
 '		</div>' +
+'' +
 '	</div>' +
 '' +
 '	<!-- TABLE -->' +
@@ -2042,34 +2081,41 @@ angular.module('keta.directives.ExtendedTable')
 '								data-ng-class="{' +
 '									sort: isSortCriteria(column),' +
 '									sortable: isSortable(column)}">' +
-'								<span data-ng-if="!isSortable(column)">' +
-'									{{headerLabelCallback(column)}}</span>' +
+'								<span data-ng-if="' +
+'									!isSortable(column) &&' +
+'									headerLabelCallback(column)">{{headerLabelCallback(column)}}</span>' +
 '								<a class="header" title="{{ getLabel(MESSAGE_KEY_PREFIX + \'_sort\') }}"' +
-'									data-ng-if="isSortable(column)"' +
+'									data-ng-if="' +
+'										isSortable(column) &&' +
+'										headerLabelCallback(column)"' +
 '									data-ng-click="sortBy(column)">{{headerLabelCallback(column)}}</a>' +
 '								<a class="sort" title="{{ getLabel(MESSAGE_KEY_PREFIX + \'_sort\') }}"' +
 '									data-ng-if="' +
 '										isSortCriteria(column) &&' +
 '										rowSortOrderAscending &&' +
-'										isSortable(column)"' +
+'										isSortable(column) &&' +
+'										headerLabelCallback(column)"' +
 '									data-ng-click="sortBy(column)"><span' +
 '									class="glyphicon glyphicon-sort-by-alphabet"></span></a>' +
 '								<a class="sort" title="{{ getLabel(MESSAGE_KEY_PREFIX + \'_sort\') }}"' +
 '								   data-ng-if="' +
 '									   isSortCriteria(column) &&' +
 '									   !rowSortOrderAscending &&' +
-'									   isSortable(column)"' +
+'									   isSortable(column) &&' +
+'									   headerLabelCallback(column)"' +
 '								   data-ng-click="sortBy(column)"><span' +
 '									class="glyphicon glyphicon-sort-by-alphabet-alt"></span></a>' +
 '								<a class="unsort" title="{{ getLabel(MESSAGE_KEY_PREFIX + \'_sort\') }}"' +
 '									data-ng-if="' +
 '										!isSortCriteria(column) &&' +
-'										headerLabelCallback(column) !== null &&' +
-'										isSortable(column)"' +
+'										isSortable(column) &&' +
+'										headerLabelCallback(column)"' +
 '									data-ng-click="sortBy(column)"><span' +
 '									class="glyphicon glyphicon-sort"></span></a>' +
 '								<a class="operation" title="{{ getLabel(MESSAGE_KEY_PREFIX + \'_remove_column\') }}"' +
-'									data-ng-if="isSwitchable(column)"' +
+'									data-ng-if="' +
+'										isSwitchable(column) &&' +
+'										headerLabelCallback(column)"' +
 '									data-ng-click="removeColumn(column)"><span' +
 '									class="glyphicon glyphicon-minus-sign"></span></a>' +
 '							</th>' +
@@ -2165,7 +2211,7 @@ angular.module('keta.directives.ExtendedTable')
 '' +
 '	<!-- PAGER -->' +
 '	<div class="row" data-ng-show="!isDisabled(COMPONENTS_PAGER) && pager !== null && pages.length > 1">' +
-'		<div class="col-xs-12 col-sm-6">' +
+'		<div class="col-xs-12">' +
 '			<div class="pager max-width">' +
 '				<div class="input-group">' +
 '					<div class="pager-buttons input-group-btn">' +
@@ -2304,8 +2350,17 @@ angular.module('keta.directives.MainMenu', [])
 				};
 
 				scope.checkExpand = function(menuEntry, $event) {
-					// close sidebar when route-links (of current route and menu-entry) are equal
-					if ($location.path() === menuEntry.link &&
+
+					var menuLink = menuEntry.link;
+
+					// strip '#' at beginning if present
+					if (angular.isString(menuLink) &&
+						menuLink.substr(0, 1) === '#') {
+						menuLink = angular.copy(menuLink.substr(1, menuLink.length - 1));
+					}
+
+					// close sidebar when route links of current route and menu entry are equal
+					if ($location.url() === menuLink &&
 						angular.isDefined(scope.configuration.toggleBroadcast)) {
 						$event.stopPropagation();
 						$rootScope.$broadcast(scope.configuration.toggleBroadcast);